add tag columns and start basics of relationships between models

+++ models/Product.js
@@ -21,7 +21,7 @@ Product.init(
       allowNull: false
     },
     price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.Decimal,
       allowNull: false,
       defaultValue: 10
